+++ src/components/plugins/workflows/expands/ExpandBySchema.php
@@ -1,24 +1,25 @@
 <?php
-namespace extas\components\plugins\workflows\expands\transitions;
+namespace extas\components\plugins\workflows\expands;
 
 use extas\components\plugins\Plugin;
 use extas\interfaces\expands\IExpand;
 use extas\interfaces\IItem;
 use extas\interfaces\repositories\IRepository;
 use extas\interfaces\stages\IStageExpand;
+use extas\interfaces\workflows\schemas\IHasSchema;
 use extas\interfaces\workflows\transitions\ITransition;
 
 /**
- * Class TransitionExpandBySchema
+ * Class ExpandBySchema
  *
  * @stage expand.index.schema
- * @package extas\components\plugins\expands\schemas
+ * @package extas\components\plugins\expands
  * @author jeyroik@gmail.com
  */
-class TransitionExpandBySchema extends Plugin implements IStageExpand
+class ExpandBySchema extends Plugin implements IStageExpand
 {
     /**
-     * @param IItem|ITransition $subject
+     * @param IItem|IHasSchema $subject
      * @param IExpand $expand
      * @return IItem
      */